# swagger.yaml
openapi: 3.0.0
info:
  title: Social Media API
  description: API documentation for the social media platform
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Enter your JWT token in the format: Bearer <your-token>
        
        Example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        
        Note: The token is obtained from the login response
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: User ID
        username:
          type: string
          description: Username
        email:
          type: string
          format: email
          description: Email address
        full_name:
          type: string
          description: Full name
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
...
paths:
...
  /api/users/follow:
    post:
      summary: Follow a user
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - following_id
              properties:
                following_id:
                  type: integer
                  description: ID of user to follow
      responses:
        '200':
          description: Follow successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/unfollow:
    delete:
      summary: Unfollow a user
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - following_id
              properties:
                following_id:
                  type: integer
                  description: ID of user to unfollow
      responses:
        '200':
          description: Unfollow successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{user_id}/followers:
    get:
      summary: Get user's followers
      tags: [Users]
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: Followers retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{user_id}/following:
    get:
      summary: Get users that the specified user is following
      tags: [Users]
      parameters:
        - in: path
          name: user_id
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: Following list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...